# server/manager/api/manager/job/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Job API
    version: 0.0.1
paths:
    /manager/api/v1/job:
        put:
            tags:
                - Job
            description: UpdateJob 更新职位信息
            operationId: Job_UpdateJob
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateJobRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UpdateJobReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        post:
            tags:
                - Job
            description: CreateJob 创建职位信息
            operationId: Job_CreateJob
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateJobRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateJobReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        delete:
            tags:
                - Job
            description: DeleteJob 删除职位信息
            operationId: Job_DeleteJob
            parameters:
                - name: id
                  in: query
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteJobReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /manager/api/v1/jobs:
        get:
            tags:
                - Job
            description: ListJob 获取职位信息列表
            operationId: Job_ListJob
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: uint32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: uint32
                - name: order
                  in: query
                  schema:
                    type: string
                - name: orderBy
                  in: query
                  schema:
                    type: string
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: name
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListJobReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        CreateJobReply:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
        CreateJobRequest:
            type: object
            properties:
                keyword:
                    type: string
                name:
                    type: string
                weight:
                    type: integer
                    format: int32
                description:
                    type: string
        DeleteJobReply:
            type: object
            properties: {}
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        ListJobReply:
            type: object
            properties:
                total:
                    type: integer
                    format: uint32
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/ListJobReply_Job'
        ListJobReply_Job:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                keyword:
                    type: string
                name:
                    type: string
                weight:
                    type: integer
                    format: int32
                description:
                    type: string
                createdAt:
                    type: integer
                    format: uint32
                updatedAt:
                    type: integer
                    format: uint32
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
        UpdateJobReply:
            type: object
            properties: {}
        UpdateJobRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                keyword:
                    type: string
                name:
                    type: string
                weight:
                    type: integer
                    format: int32
                description:
                    type: string
tags:
    - name: Job